@@ -50,7 +50,7 @@ RSpec.describe Msf::Exploit::AutoTarget do
   describe '#auto_target?' do
     it 'should return true if the automatic target is selected' do
       host_addr = '192.168.1.5'
-      host_obj  = FactoryGirl.create(:mdm_host, address: host_addr )
+      host_obj  = FactoryBot.create(:mdm_host, address: host_addr )
       windows_exploit.datastore['TARGET'] = 0
       windows_exploit.datastore['WORKSPACE'] = host_obj.workspace.name
       windows_exploit.datastore['RHOST'] = host_addr
@@ -71,7 +71,7 @@ RSpec.describe Msf::Exploit::AutoTarget do
   context 'finding the target host' do
     it 'should return a matching Mdm::host if there is one' do
       host_addr = '192.168.1.5'
-      host_obj  = FactoryGirl.create(:mdm_host, address: host_addr )
+      host_obj  = FactoryBot.create(:mdm_host, address: host_addr )
       windows_exploit.datastore['WORKSPACE'] = host_obj.workspace.name
       windows_exploit.datastore['RHOST'] = host_addr
       expect(windows_exploit.auto_target_host).to eq host_obj
@@ -84,12 +84,12 @@ RSpec.describe Msf::Exploit::AutoTarget do
   end
 
   context 'filtering targets' do
-    let(:windows_xp_host) { FactoryGirl.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP' ) }
-    let(:windows_xp_sp1_host) { FactoryGirl.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP', os_sp: 'SP1' ) }
-    let(:windows_xp_sp2_host) { FactoryGirl.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP', os_sp: 'SP2' ) }
-    let(:windows_xp_sp3_host) { FactoryGirl.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP', os_sp: 'SP3' ) }
-    let(:windows_7_host) { FactoryGirl.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows 7' ) }
-    let(:unknown_host) { FactoryGirl.create(:mdm_host, address: '192.168.172.150', os_family: nil ) }
+    let(:windows_xp_host) { FactoryBot.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP' ) }
+    let(:windows_xp_sp1_host) { FactoryBot.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP', os_sp: 'SP1' ) }
+    let(:windows_xp_sp2_host) { FactoryBot.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP', os_sp: 'SP2' ) }
+    let(:windows_xp_sp3_host) { FactoryBot.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows XP', os_sp: 'SP3' ) }
+    let(:windows_7_host) { FactoryBot.create(:mdm_host, address: '192.168.172.150', os_family: 'Windows', os_name: 'Windows 7' ) }
+    let(:unknown_host) { FactoryBot.create(:mdm_host, address: '192.168.172.150', os_family: nil ) }
     let(:potential_targets) { windows_exploit.filter_by_os_family(windows_xp_host) }
     let(:xp_targets) { windows_exploit.filter_by_os_name(potential_targets,windows_xp_host) }
 
