@@ -11,15 +11,15 @@ RSpec.shared_examples_for 'Msf::DBManager::ExploitAttempt' do
 
     let(:run) do
       match
-      FactoryGirl.create(:automatic_exploitation_run, user: workspace.owner,workspace:workspace, match_set_id: match_set.id)
+      FactoryBot.create(:automatic_exploitation_run, user: workspace.owner,workspace:workspace, match_set_id: match_set.id)
     end
 
     let(:match_set) do
-      FactoryGirl.create(:automatic_exploitation_match_set, user: workspace.owner,workspace:workspace)
+      FactoryBot.create(:automatic_exploitation_match_set, user: workspace.owner,workspace:workspace)
     end
 
     let(:match) do
-      FactoryGirl.create(:automatic_exploitation_match,
+      FactoryBot.create(:automatic_exploitation_match,
                          match_set_id: match_set.id,
                          matchable_id:vuln_with_match.id,
                          matchable_type: "Mdm::Vuln"
@@ -27,19 +27,19 @@ RSpec.shared_examples_for 'Msf::DBManager::ExploitAttempt' do
     end
 
     let(:vuln_with_match) do
-      FactoryGirl.create(:mdm_vuln)
+      FactoryBot.create(:mdm_vuln)
     end
 
     let(:host) do
-      FactoryGirl.create(:mdm_host, workspace:workspace)
+      FactoryBot.create(:mdm_host, workspace:workspace)
     end
 
     let(:workspace) do
-      FactoryGirl.create(:mdm_workspace)
+      FactoryBot.create(:mdm_workspace)
     end
 
     let(:refs) do
-      [ FactoryGirl.create(:mdm_ref) ]
+      [ FactoryBot.create(:mdm_ref) ]
     end
 
     context "with a run" do
@@ -137,7 +137,7 @@ RSpec.shared_examples_for 'Msf::DBManager::ExploitAttempt' do
 
     context "without a run" do
       let(:vuln) do
-        FactoryGirl.create(:mdm_vuln)
+        FactoryBot.create(:mdm_vuln)
       end
 
       let(:opts) do
@@ -226,20 +226,20 @@ RSpec.shared_examples_for 'Msf::DBManager::ExploitAttempt' do
     end
 
     let(:session_id) do
-      FactoryGirl.create(:session, host: host).id
+      FactoryBot.create(:session, host: host).id
     end
 
     let(:run) do
       match
-      FactoryGirl.create(:automatic_exploitation_run, user: workspace.owner,workspace:workspace, match_set_id: match_set.id)
+      FactoryBot.create(:automatic_exploitation_run, user: workspace.owner,workspace:workspace, match_set_id: match_set.id)
     end
 
     let(:match_set) do
-      FactoryGirl.create(:automatic_exploitation_match_set, user: workspace.owner,workspace:workspace)
+      FactoryBot.create(:automatic_exploitation_match_set, user: workspace.owner,workspace:workspace)
     end
 
     let(:match) do
-      FactoryGirl.create(:automatic_exploitation_match,
+      FactoryBot.create(:automatic_exploitation_match,
                          match_set_id: match_set.id,
                          matchable_id:vuln_with_match.id,
                          matchable_type: "Mdm::Vuln"
@@ -247,19 +247,19 @@ RSpec.shared_examples_for 'Msf::DBManager::ExploitAttempt' do
     end
 
     let(:vuln_with_match) do
-      FactoryGirl.create(:mdm_vuln)
+      FactoryBot.create(:mdm_vuln)
     end
 
     let(:host) do
-      FactoryGirl.create(:mdm_host, workspace:workspace)
+      FactoryBot.create(:mdm_host, workspace:workspace)
     end
 
     let(:workspace) do
-      FactoryGirl.create(:mdm_workspace)
+      FactoryBot.create(:mdm_workspace)
     end
 
     let(:refs) do
-      [ FactoryGirl.create(:mdm_ref) ]
+      [ FactoryBot.create(:mdm_ref) ]
     end
 
     context "with a run" do
@@ -357,7 +357,7 @@ RSpec.shared_examples_for 'Msf::DBManager::ExploitAttempt' do
 
     context "without a run" do
       let(:vuln) do
-        FactoryGirl.create(:mdm_vuln)
+        FactoryBot.create(:mdm_vuln)
       end
 
       let(:opts) do
