@@ -33,7 +33,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
           end
 
           let(:host) do
-            FactoryGirl.create(:mdm_host, :workspace => session_workspace)
+            FactoryBot.create(:mdm_host, :workspace => session_workspace)
           end
 
           let(:module_instance) do
@@ -50,7 +50,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
           end
 
           let(:options_workspace) do
-            FactoryGirl.create(:mdm_workspace)
+            FactoryBot.create(:mdm_workspace)
           end
 
           let(:parent_module_fullname) do
@@ -95,7 +95,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
           end
 
           let(:session_workspace) do
-            FactoryGirl.create(:mdm_workspace)
+            FactoryBot.create(:mdm_workspace)
           end
 
           before(:example) do
@@ -115,7 +115,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
                 }
             )
 
-            FactoryGirl.create(
+            FactoryBot.create(
                 :mdm_module_detail,
                 :fullname => parent_module_fullname,
                 :name => parent_module_name
@@ -128,11 +128,11 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
             end
 
             let(:match_set) do
-              FactoryGirl.create(:automatic_exploitation_match_set, user: session_workspace.owner,workspace:session_workspace)
+              FactoryBot.create(:automatic_exploitation_match_set, user: session_workspace.owner,workspace:session_workspace)
             end
 
             let(:run) do
-              FactoryGirl.create(:automatic_exploitation_run, workspace: session_workspace, match_set_id: match_set.id)
+              FactoryBot.create(:automatic_exploitation_run, workspace: session_workspace, match_set_id: match_set.id)
             end
 
             let(:user_data) do
@@ -172,24 +172,10 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
             end
 
             context 'with workspace from either :workspace or session' do
-              it 'should pass normalized host from session as :host to #find_or_create_host' do
-                normalized_host = double('Normalized Host')
-                expect(db_manager).to receive(:normalize_host).with(session).and_return(normalized_host)
-                # stub report_vuln so its use of find_or_create_host and normalize_host doesn't interfere.
-                expect(db_manager).to receive(:report_vuln)
-
-                expect(db_manager).to receive(:find_or_create_host).with(
-                  hash_including(
-                    :host => normalized_host
-                  )
-                ).and_return(host)
-
-                report_session
-              end
 
               context 'with session responds to arch' do
                 let(:arch) do
-                  FactoryGirl.generate :mdm_host_arch
+                  FactoryBot.generate :mdm_host_arch
                 end
 
                 before(:example) do
@@ -320,7 +306,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
                     end
 
                     let(:service) do
-                      FactoryGirl.create(
+                      FactoryBot.create(
                         :mdm_service,
                         :host => host
                       )
@@ -515,24 +501,10 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
             end
 
             context 'with workspace from either :workspace or session' do
-              it 'should pass normalized host from session as :host to #find_or_create_host' do
-                normalized_host = double('Normalized Host')
-                allow(db_manager).to receive(:normalize_host).with(session).and_return(normalized_host)
-                # stub report_vuln so its use of find_or_create_host and normalize_host doesn't interfere.
-                allow(db_manager).to receive(:report_vuln)
-
-                expect(db_manager).to receive(:find_or_create_host).with(
-                    hash_including(
-                        :host => normalized_host
-                    )
-                ).and_return(host)
-
-                report_session
-              end
 
               context 'with session responds to arch' do
                 let(:arch) do
-                  FactoryGirl.generate :mdm_host_arch
+                  FactoryBot.generate :mdm_host_arch
                 end
 
                 before(:example) do
@@ -663,7 +635,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
                     end
 
                     let(:service) do
-                      FactoryGirl.create(
+                      FactoryBot.create(
                           :mdm_service,
                           :host => host
                       )
@@ -846,7 +818,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
 
           context 'with Mdm::Host' do
             let(:host) do
-              FactoryGirl.create(:mdm_host)
+              FactoryBot.create(:mdm_host)
             end
 
             context 'created Mdm::Session' do
@@ -951,7 +923,7 @@ RSpec.shared_examples_for 'Msf::DBManager::Session' do
 
               context 'with :routes' do
                 let(:routes) do
-                  FactoryGirl.build_list(
+                  FactoryBot.build_list(
                       :mdm_route,
                       1,
                       :session => nil
