@@ -9,7 +9,6 @@ require 'rex/socket'
 #
 # Project
 #
-
 require 'metasploit/framework/require'
 require 'msf/base/config'
 require 'msf/core'
@@ -17,6 +16,8 @@ require 'msf/core/database_event'
 require 'msf/core/db_import_error'
 require 'msf/core/host_state'
 require 'msf/core/service_state'
+require 'metasploit/framework/data_service'
+
 
 # The db module provides persistent storage and events. This class should be instantiated LAST
 # as the active_suppport library overrides Kernel.require, slowing down all future code loads.
@@ -31,6 +32,7 @@ class Msf::DBManager
   autoload :Client, 'msf/core/db_manager/client'
   autoload :Connection, 'msf/core/db_manager/connection'
   autoload :Cred, 'msf/core/db_manager/cred'
+  autoload :DbExport, 'msf/core/db_manager/db_export'
   autoload :Event, 'msf/core/db_manager/event'
   autoload :ExploitAttempt, 'msf/core/db_manager/exploit_attempt'
   autoload :ExploitedHost, 'msf/core/db_manager/exploited_host'
@@ -60,10 +62,14 @@ class Msf::DBManager
 
   optionally_include_metasploit_credential_creation
 
+  # Interface must be included first
+  include Metasploit::Framework::DataService
+
   include Msf::DBManager::Adapter
   include Msf::DBManager::Client
   include Msf::DBManager::Connection
   include Msf::DBManager::Cred
+  include Msf::DBManager::DbExport
   include Msf::DBManager::Event
   include Msf::DBManager::ExploitAttempt
   include Msf::DBManager::ExploitedHost
@@ -92,6 +98,10 @@ class Msf::DBManager
 
   # Provides :framework and other accessors
   include Msf::Framework::Offspring
+
+  def name
+    'local_db_service'
+  end
 
   #
   # Attributes
@@ -122,7 +132,7 @@ class Msf::DBManager
       return
     end
 
-    initialize_database_support
+    return initialize_database_support
   end
 
   #
@@ -162,5 +172,62 @@ class Msf::DBManager
     initialize_adapter
 
     true
+  end
+
+  def init_db(opts)
+
+    init_success = false
+
+    # Append any migration paths necessary to bring the database online
+    if opts['DatabaseMigrationPaths']
+      opts['DatabaseMigrationPaths'].each do |migrations_path|
+        ActiveRecord::Migrator.migrations_paths << migrations_path
+      end
+    end
+
+    if connection_established?
+      after_establish_connection
+    else
+      configuration_pathname = Metasploit::Framework::Database.configurations_pathname(path: opts['DatabaseYAML'])
+
+      unless configuration_pathname.nil?
+        if configuration_pathname.readable?
+          dbinfo = YAML.load_file(configuration_pathname) || {}
+          dbenv  = opts['DatabaseEnv'] || Rails.env
+          db     = dbinfo[dbenv]
+        else
+          elog("Warning, #{configuration_pathname} is not readable. Try running as root or chmod.")
+        end
+
+        if not db
+          elog("No database definition for environment #{dbenv}")
+        else
+          init_success = connect(db)
+        end
+      end
+    end
+
+    # framework.db.active will be true if after_establish_connection ran directly when connection_established? was
+    # already true or if framework.db.connect called after_establish_connection.
+    if !! error
+      if error.to_s =~ /RubyGem version.*pg.*0\.11/i
+        elog("***")
+        elog("*")
+        elog("* Metasploit now requires version 0.11 or higher of the 'pg' gem for database support")
+        elog("* There a three ways to accomplish this upgrade:")
+        elog("* 1. If you run Metasploit with your system ruby, simply upgrade the gem:")
+        elog("*    $ rvmsudo gem install pg ")
+        elog("* 2. Use the Community Edition web interface to apply a Software Update")
+        elog("* 3. Uninstall, download the latest version, and reinstall Metasploit")
+        elog("*")
+        elog("***")
+        elog("")
+        elog("")
+      end
+
+      elog("Failed to connect to the database: #{error}")
+    end
+
+    return init_success
   end
 end
