@@ -1,4 +1,13 @@
 module Msf::DBManager::Session
+
+  def get_all_sessions()
+    return if not active
+    ::ActiveRecord::Base.connection_pool.with_connection {
+      ::Mdm::Session.all
+    }
+  end
+
+
   # Returns a session based on opened_time, host address, and workspace
   # (or returns nil)
   def get_session(opts)
@@ -41,7 +50,7 @@ module Msf::DBManager::Session
   #     created, but the Mdm::Host will have been.   (There is no transaction
   #       to rollback the Mdm::Host creation.)
   #   @see #find_or_create_host
-  #   @see #normalize_host
+  #   @see #Msf::Util::Host.normalize_host
   #   @see #report_exploit_success
   #   @see #report_vuln
   #
@@ -101,7 +110,74 @@ module Msf::DBManager::Session
   }
   end
 
+  def report_session_host_dto(host_dto)
+    host_dto[:host] = get_host(host_dto)
+    create_mdm_session_from_host(host_dto)
+  end
+
+  def report_session_dto(session_dto)
+    return if not active
+
+    ::ActiveRecord::Base.connection_pool.with_connection {
+      workspace = find_workspace(session_dto[:workspace])
+      host_data = session_dto[:host_data]
+      h_opts = {}
+      h_opts[:host]      = host_data[:host]
+      h_opts[:arch]      = host_data[:arch]
+      h_opts[:workspace] = workspace
+      host = find_or_create_host(h_opts)
+
+      session_data = session_dto[:session_data]
+      sess_data = {
+          datastore: session_data[:datastore],
+          desc: session_data[:desc],
+          host_id: host.id,
+          last_seen: session_dto[:time_stamp],
+          local_id: session_data[:local_id],
+          opened_at: session_dto[:time_stamp],
+          platform: session_data[:platform],
+          port: session_data[:port],
+          routes: [],
+          stype: session_data[:stype],
+          via_exploit: session_data[:via_exploit],
+          via_payload: session_data[:via_payload],
+      }
+
+      if sess_data[:via_exploit] == 'exploit/multi/handler' and sess_data[:datastore] and sess_data[:datastore]['ParentModule']
+        sess_data[:via_exploit] = sess_data[:datastore]['ParentModule']
+      end
+
+      session_db_record = ::Mdm::Session.create!(sess_data)
+
+      # TODO: Figure out task stuff
+
+
+      if sess_data[:via_exploit]
+        # This is a live session, we know the host is vulnerable to something.
+        infer_vuln_from_session_dto(session_dto, session_db_record, workspace)
+      end
+      session_db_record
+    }
+  end
+
+  # Clean out any stale sessions that have been orphaned by a dead framework instance.
+  # @param last_seen_interval [Integer] interval, in seconds, open sessions are marked as alive
+  def remove_stale_sessions(last_seen_interval)
+    return unless active
+
+    ::ActiveRecord::Base.connection_pool.with_connection {
+      ::Mdm::Session.where(closed_at: nil).each do |db_session|
+        next unless db_session.last_seen.nil? or ((Time.now.utc - db_session.last_seen) > (2 * last_seen_interval))
+        db_session.closed_at    = db_session.last_seen || Time.now.utc
+        db_session.close_reason = "Orphaned"
+        db_session.save
+      end
+    }
+  end
+
+  #########
   protected
+  #########
 
   # @param session [Msf::Session] A session with a db_record Msf::Session#db_record
   # @param wspace [Mdm::Workspace]
@@ -117,10 +193,14 @@ module Msf::DBManager::Session
         mod_fullname = session.via_exploit
       end
       mod_detail = ::Mdm::Module::Detail.find_by_fullname(mod_fullname)
+
       if mod_detail.nil?
         # Then the cache isn't built yet, take the hit for instantiating the
         # module
         mod_detail = framework.modules.create(mod_fullname)
+        refs = mod_detail.references
+      else
+        refs = mod_detail.refs
       end
       mod_name = mod_detail.name
 
@@ -129,7 +209,7 @@ module Msf::DBManager::Session
         host: host,
         info: "Exploited by #{mod_fullname} to create Session #{s.id}",
         name: mod_name,
-        refs: mod_detail.refs.map(&:name),
+        refs: refs,
         workspace: wspace,
       }
 
@@ -143,7 +223,7 @@ module Msf::DBManager::Session
       attempt_info = {
         host: host,
         module: mod_fullname,
-        refs: mod_detail.refs,
+        refs: refs,
         service: service,
         session_id: s.id,
         timestamp: Time.now.utc,
@@ -166,7 +246,7 @@ module Msf::DBManager::Session
 
       wspace = opts[:workspace] || find_workspace(session.workspace)
       h_opts = { }
-      h_opts[:host]      = normalize_host(session)
+      h_opts[:host]      = Msf::Util::Host.normalize_host(session)
       h_opts[:arch]      = session.arch if session.respond_to?(:arch) and session.arch
       h_opts[:workspace] = wspace
       host = find_or_create_host(h_opts)
@@ -230,4 +310,46 @@ module Msf::DBManager::Session
     }
   end
 
+  def infer_vuln_from_session_dto(session_dto, session_db_record, workspace)
+    ::ActiveRecord::Base.connection_pool.with_connection {
+
+      vuln_info_dto = session_dto[:vuln_info]
+      host = session_db_record.host
+      mod_name = vuln_info_dto[:mod_name] || 'unknown'
+      refs = vuln_info_dto[:mod_references] || []
+
+      vuln_info = {
+          exploited_at: session_dto[:time_stamp],
+          host: host,
+          info: "Exploited by #{vuln_info_dto[:mod_fullname]} to create Session #{session_db_record.id}",
+          name: mod_name,
+          refs: refs,
+          workspace: workspace,
+      }
+
+      port    = vuln_info_dto[:remote_port]
+      service = (port ? host.services.find_by_port(port.to_i) : nil)
+
+      vuln_info[:service] = service if service
+
+      vuln = framework.db.report_vuln(vuln_info)
+
+      attempt_info = {
+          host: host,
+          module: vuln_info_dto[:mod_fullname],
+          refs: refs,
+          service: service,
+          session_id: session_db_record.id,
+          timestamp: session_dto[:time_stamp],
+          username: vuln_info_dto[:username],
+          vuln: vuln,
+          workspace: workspace,
+          run_id: vuln_info_dto[:run_id]
+      }
+
+      framework.db.report_exploit_success(attempt_info)
+
+      vuln
+    }
+  end
 end
