@@ -58,7 +58,7 @@ class Framework
   require 'msf/core/module_manager'
   require 'msf/core/session_manager'
   require 'msf/core/plugin_manager'
-  require 'msf/core/db_manager'
+  require 'metasploit/framework/data_service/proxy/core'
   require 'msf/core/event_dispatcher'
   require 'rex/json_hash_file'
   require 'msf/core/cert_provider'
@@ -193,13 +193,13 @@ class Framework
   #
   attr_reader   :browser_profiles
 
-  # The framework instance's db manager. The db manager
-  # maintains the database db and handles db events
   #
-  # @return [Msf::DBManager]
+  # The framework instance's data service proxy
+  #
+  # @return [Metasploit::Framework::DataService::DataProxy]
   def db
     synchronize {
-      @db ||= Msf::DBManager.new(self, options)
+      @db ||= get_db
     }
   end
 
@@ -268,6 +268,19 @@ protected
   attr_writer   :db # :nodoc:
   attr_writer   :uuid_db # :nodoc:
   attr_writer   :browser_profiles # :nodoc:
+
+  private
+
+  def get_db
+    if !options['DisableDatabase']
+      db_manager = Msf::DBManager.new(self)
+      db_manager.init_db(options)
+      options[:db_manager] = db_manager
+    end
+
+    Metasploit::Framework::DataService::DataProxy.new(options)
+  end
+
 end
 
 class FrameworkEventSubscriber
@@ -326,7 +339,7 @@ class FrameworkEventSubscriber
   ##
   # :category: ::Msf::UiEventSubscriber implementors
   def on_ui_command(command)
-    if framework.db.active
+    if (framework.db and framework.db.active)
       report_event(:name => "ui_command", :info => {:command => command})
     end
   end
@@ -334,7 +347,7 @@ class FrameworkEventSubscriber
   ##
   # :category: ::Msf::UiEventSubscriber implementors
   def on_ui_stop()
-    if framework.db.active
+    if (framework.db and framework.db.active)
       report_event(:name => "ui_stop")
     end
   end
