@@ -19,6 +19,10 @@ module Msf::Modules::Metadata::Store
     load_metadata
   end
 
+  #######
+  private
+  #######
+
   #
   # Update the module meta cache disk store
   #
@@ -31,10 +35,6 @@ module Msf::Modules::Metadata::Store
       elog("Unable to update metadata store: #{e.message}")
     end
   end
-
-  #######
-  private
-  #######
 
   def load_metadata
     begin
