@@ -50,18 +50,11 @@ module Msf
       Msf::OptPort.new(__method__.to_s, [ required, desc, default ])
     end
 
-    def self.ssl_supported_options
-      @m ||= ['Auto', 'TLS'] + OpenSSL::SSL::SSLContext::METHODS \
-             .select{|m| !m.to_s.include?('client') && !m.to_s.include?('server')} \
-             .select{|m| OpenSSL::SSL::SSLContext.new(m) && true rescue false} \
-             .map{|m| m.to_s.sub(/v/, '').sub('_', '.')}
-    end
-
     # @return [OptEnum]
     def self.SSLVersion
       Msf::OptEnum.new('SSLVersion',
         'Specify the version of SSL/TLS to be used (Auto, TLS and SSL23 are auto-negotiate)',
-        enums: self.ssl_supported_options
+        enums: Rex::Socket::SslTcp.supported_ssl_methods
       )
     end
 
