@@ -108,7 +108,7 @@ class ThreadManager < Array
           elog("Call Stack\n#{e.backtrace.join("\n")}")
           raise e
         ensure
-          if framework.db and framework.db.active
+          if framework.db && framework.db.active && framework.db.is_local?
             # NOTE: despite the Deprecation Warning's advice, this should *NOT*
             # be ActiveRecord::Base.connection.close which causes unrelated
             # threads to raise ActiveRecord::StatementInvalid exceptions at
