@@ -290,6 +290,7 @@ class Creds
     port_ranges   = []
     svcs          = []
     rhosts        = []
+    opts          = {}
 
     set_rhosts = false
 
@@ -314,6 +315,7 @@ class Creds
         end
       when "-t","--type"
         ptype = args.shift
+        opts[:ptype] = ptype
         if (!ptype)
           print_error("Argument required for -t")
           return
@@ -325,14 +327,17 @@ class Creds
           return
         end
         svcs = service.split(/[\s]*,[\s]*/)
+        opts[:svcs] = svcs
       when "-P","--password"
         pass = args.shift
+        opts[:pass] = pass
         if (!pass)
           print_error("Argument required for -P")
           return
         end
       when "-u","--user"
         user = args.shift
+        opts[:user] = user
         if (!user)
           print_error("Argument required for -u")
           return
@@ -343,6 +348,7 @@ class Creds
         set_rhosts = true
       when '-O', '--origins'
         hosts = args.shift
+        opts[:hosts] = hosts
         if !hosts
           print_error("Argument required for -O")
           return
@@ -350,6 +356,7 @@ class Creds
         arg_host_range(hosts, origin_ranges)
       when '-S', '--search-term'
         search_term = args.shift
+        opts[:search_term] = search_term
       else
         # Anything that wasn't an option is a host to search for
         unless (arg_host_range(arg, host_ranges))
@@ -374,6 +381,8 @@ class Creds
              end
     end
 
+    opts[:type] = type if type
+
     # normalize
     ports = port_ranges.flatten.uniq
     svcs.flatten!
@@ -384,120 +393,98 @@ class Creds
     }
 
     tbl = Rex::Text::Table.new(tbl_opts)
+    opts[:wspace] = framework.db.workspace
+    query = framework.db.creds(opts)
 
-    ::ActiveRecord::Base.connection_pool.with_connection {
-      query = Metasploit::Credential::Core.where( workspace_id: framework.db.workspace )
-      query = query.includes(:private, :public, :logins).references(:private, :public, :logins)
-      query = query.includes(logins: [ :service, { service: :host } ])
+    query.each do |core|
 
-      if type.present?
-        query = query.where(metasploit_credential_privates: { type: type })
+      # Exclude non-blank username creds if that's what we're after
+      if user == "" && core.public && !(core.public.username.blank?)
+        next
       end
 
-      if svcs.present?
-        query = query.where(Mdm::Service[:name].in(svcs))
+      # Exclude non-blank password creds if that's what we're after
+      if pass == "" && core.private && !(core.private.data.blank?)
+        next
       end
 
-      if ports.present?
-        query = query.where(Mdm::Service[:port].in(ports))
+      origin = ''
+      if core.origin.kind_of?(Metasploit::Credential::Origin::Service)
+        origin = core.origin.service.host.address
+      elsif core.origin.kind_of?(Metasploit::Credential::Origin::Session)
+        origin = core.origin.session.host.address
       end
 
-      if user.present?
-        # If we have a user regex, only include those that match
-        query = query.where('"metasploit_credential_publics"."username" ~* ?', user)
+      if !origin.empty? && origin_ranges.present? && !origin_ranges.any? {|range| range.include?(origin) }
+        next
       end
 
-      if pass.present?
-        # If we have a password regex, only include those that match
-        query = query.where('"metasploit_credential_privates"."data" ~* ?', pass)
-      end
+      if core.logins.empty? && origin_ranges.empty?
+        public_val = core.public ? core.public.username : ""
+        private_val = core.private ? core.private.data : ""
+        realm_val = core.realm ? core.realm.value : ""
+        human_val = core.private ? core.private.class.model_name.human : ""
 
-      if host_ranges.any? || ports.any? || svcs.any?
-        # Only find Cores that have non-zero Logins if the user specified a
-        # filter based on host, port, or service name
-        query = query.where(Metasploit::Credential::Login[:id].not_eq(nil))
-      end
-
-      query.find_each do |core|
-
-        # Exclude non-blank username creds if that's what we're after
-        if user == "" && core.public && !(core.public.username.blank?)
-          next
-        end
-
-        # Exclude non-blank password creds if that's what we're after
-        if pass == "" && core.private && !(core.private.data.blank?)
-          next
-        end
-
-        origin = ''
-        if core.origin.kind_of?(Metasploit::Credential::Origin::Service)
-          origin = core.origin.service.host.address
-        elsif core.origin.kind_of?(Metasploit::Credential::Origin::Session)
-          origin = core.origin.session.host.address
-        end
-
-        if !origin.empty? && origin_ranges.present? && !origin_ranges.any? {|range| range.include?(origin) }
-          next
-        end
-
-        if core.logins.empty? && origin_ranges.empty?
-          tbl << [
-            "", # host
-            "", # cred
-            "", # service
-            core.public,
-            core.private,
-            core.realm,
-            core.private ? core.private.class.model_name.human : "",
-          ]
-        else
-          core.logins.each do |login|
-            # If none of this Core's associated Logins is for a host within
-            # the user-supplied RangeWalker, then we don't have any reason to
-            # print it out. However, we treat the absence of ranges as meaning
-            # all hosts.
-            if host_ranges.present? && !host_ranges.any? { |range| range.include?(login.service.host.address) }
-              next
-            end
-
-            row = [ login.service.host.address ]
-            row << origin
-            rhosts << login.service.host.address
-            if login.service.name.present?
-              row << "#{login.service.port}/#{login.service.proto} (#{login.service.name})"
-            else
-              row << "#{login.service.port}/#{login.service.proto}"
-            end
-
-            row += [
-              core.public,
-              core.private,
-              core.realm,
-              core.private ? core.private.class.model_name.human : "",
-            ]
-            tbl << row
-          end
-        end
-        if mode == :delete
-          core.destroy
-          delete_count += 1
-        end
-      end
-
-      if output_file.nil?
-        print_line(tbl.to_s)
+        tbl << [
+          "", # host
+          "", # cred
+          "", # service
+          public_val,
+          private_val,
+          realm_val,
+          human_val
+        ]
       else
-        # create the output file
-        ::File.open(output_file, "wb") { |f| f.write(tbl.to_csv) }
-        print_status("Wrote creds to #{output_file}")
-      end
+        core.logins.each do |login|
+          # If none of this Core's associated Logins is for a host within
+          # the user-supplied RangeWalker, then we don't have any reason to
+          # print it out. However, we treat the absence of ranges as meaning
+          # all hosts.
+          if host_ranges.present? && !host_ranges.any? { |range| range.include?(login.service.host.address) }
+            next
+          end
 
-      # Finally, handle the case where the user wants the resulting list
-      # of hosts to go into RHOSTS.
-      set_rhosts_from_addrs(rhosts.uniq) if set_rhosts
-      print_status("Deleted #{delete_count} creds") if delete_count > 0
-    }
+          row = [ login.service.host.address ]
+          row << origin
+          rhosts << login.service.host.address
+          if login.service.name.present?
+            row << "#{login.service.port}/#{login.service.proto} (#{login.service.name})"
+          else
+            row << "#{login.service.port}/#{login.service.proto}"
+          end
+
+          public_val = core.public ? core.public.username : ""
+          private_val = core.private ? core.private.data : ""
+          realm_val = core.realm ? core.realm.value : ""
+          human_val = core.private ? core.private.class.model_name.human : ""
+
+          row += [
+            public_val,
+            private_val,
+            realm_val,
+            human_val
+          ]
+          tbl << row
+        end
+      end
+      if mode == :delete
+        core.destroy
+        delete_count += 1
+      end
+    end
+
+    if output_file.nil?
+      print_line(tbl.to_s)
+    else
+      # create the output file
+      ::File.open(output_file, "wb") { |f| f.write(tbl.to_csv) }
+      print_status("Wrote creds to #{output_file}")
+    end
+
+    # Finally, handle the case where the user wants the resulting list
+    # of hosts to go into RHOSTS.
+    set_rhosts_from_addrs(rhosts.uniq) if set_rhosts
+    print_status("Deleted #{delete_count} creds") if delete_count > 0
   end
 
   def cmd_creds_tabs(str, words)
