@@ -7,6 +7,7 @@ class MetasploitModule < Msf::Auxiliary
   include Msf::Exploit::Remote::DCERPC
   include Msf::Exploit::Remote::SMB::Client
   include Msf::Exploit::Remote::SMB::Client::Authenticated
+  include Msf::Exploit::Remote::SMB::Client::PipeAuditor
 
   include Msf::Auxiliary::Scanner
   include Msf::Auxiliary::Report
@@ -51,8 +52,9 @@ class MetasploitModule < Msf::Auxiliary
 
     register_options(
       [
-        OptBool.new('CHECK_DOPU', [true, 'Check for DOUBLEPULSAR on vulnerable hosts', true]),
-        OptBool.new('CHECK_ARCH', [true, 'Check for architecture on vulnerable hosts', true])
+        OptBool.new('CHECK_DOPU', [false, 'Check for DOUBLEPULSAR on vulnerable hosts', true]),
+        OptBool.new('CHECK_ARCH', [false, 'Check for architecture on vulnerable hosts', true]),
+        OptBool.new('CHECK_PIPE', [false, 'Check for named pipe on vulnerable hosts', false])
       ])
   end
 
@@ -112,6 +114,23 @@ class MetasploitModule < Msf::Auxiliary
               info: "MultiPlexID += 0x10 on Trans2 request - Arch: #{arch}, XOR Key: 0x#{xor_key}"
             )
           end
+        end
+
+        if datastore['CHECK_PIPE']
+          pipe_name, _ = check_named_pipes(return_first: true)
+
+          return unless pipe_name
+
+          print_good("Named pipe found: #{pipe_name}")
+
+          report_note(
+            host:  ip,
+            port:  rport,
+            proto: 'tcp',
+            sname: 'smb',
+            type:  'MS17-010 Named Pipe',
+            data:  pipe_name
+          )
         end
       elsif status == "STATUS_ACCESS_DENIED" or status == "STATUS_INVALID_HANDLE"
         # STATUS_ACCESS_DENIED (Windows 10) and STATUS_INVALID_HANDLE (others)
