@@ -18,7 +18,7 @@ class MetasploitModule < Msf::Auxiliary
       },
       'Author' => 'Andrew Morris <andrew[at]morris.guru>',
       'References' => [
-        ['URL', 'https://cultofthedyingsun.wordpress.com/2014/09/12/death-by-magick-number-fingerprinting-kippo-2014/'],
+        ['URL', 'https://www.obscurechannel.com/x42/magicknumber.html'],
         ['URL', 'http://morris.guru/detecting-kippo-ssh-honeypots/']
       ],
       'License' => MSF_LICENSE
