@@ -6,12 +6,12 @@ RSpec.shared_context 'Msf::DBManager' do
   end
 
   let(:db_manager) do
-    framework.db
+    framework.db.get_data_service
   end
 
   before(:example) do
     # already connected due to use_transactional_fixtures, but need some of the side-effects of #connect
-    framework.db.workspace = framework.db.default_workspace
+    db_manager.workspace = db_manager.default_workspace
     allow(db_manager).to receive(:active).and_return(active)
   end
 end
